# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/open-apis/core/jwt/token/tenant_token": {
            "put": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jwt模块"
                ],
                "summary": "刷新TenantAccessToken(TenantRefreshToken)",
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jwt模块"
                ],
                "summary": "生成TenantAccessToken(UserAccessToken)",
                "parameters": [
                    {
                        "description": "请求参数体",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/jwt.EnterprisePk"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/open-apis/core/jwt/token/user_token": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jwt模块"
                ],
                "summary": "后台扫码登录\u003e获取用户UserAccessToken",
                "parameters": [
                    {
                        "type": "string",
                        "description": "key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jwt模块"
                ],
                "summary": "刷新UserAccessToken（UserRefreshToken）",
                "responses": {}
            }
        },
...
        "/open-apis/core/rbac/resource/interface": {
            "get": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "获取接口信息(UserAccessToken)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "接口方法",
                        "name": "interface_way",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "接口路径",
                        "name": "interface_url",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "接口名",
                        "name": "interface_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "接口标识符",
                        "name": "interface_key",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "修改接口信息(UserAccessToken)",
                "parameters": [
                    {
                        "description": "请求参数体",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rbac.UpdateInterfaceResourceReq"
                        }
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "新增接口信息(UserAccessToken)",
                "parameters": [
                    {
                        "description": "请求参数体",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rbac.CreateInterfaceResourceReq"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "删除接口信息(UserAccessToken)",
                "parameters": [
                    {
                        "description": "请求参数体",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rbac.DeleteReq"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "jwt.EnterprisePk": {
            "type": "object",
            "properties": {
                "enterprise_pk": {
                    "type": "string"
                }
            }
        },
...
        "rbac.CreateInterfaceResourceReq": {
            "type": "object",
            "properties": {
                "interface_key": {
                    "type": "string"
                },
                "interface_name": {
                    "type": "string"
                },
                "interface_url": {
                    "type": "string"
                },
                "interface_way": {
                    "type": "string"
                }
            }
        },
...
        "rbac.DeleteReq": {
            "type": "object",
            "properties": {
                "pk": {
                    "type": "string"
                }
            }
        },
...
        "rbac.UpdateInterfaceResourceReq": {
            "type": "object",
            "properties": {
                "interface_key": {
                    "type": "string"
                },
                "interface_name": {
                    "type": "string"
                },
                "interface_url": {
                    "type": "string"
                },
                "interface_way": {
                    "type": "string"
                },
                "pk": {
                    "type": "string"
                }
            }
        },
...
    }
}